refactoring + add aria
